Update skeleton for multiplechoice automator.py
Essentially the same as main (openai api); however utilises Gemini rather than OpenAi. Main differences is that Flash 2.0 is free up to a limited request rate per minute, and the file size is reduced by about 0.01 KB (if you care). Note how this branch is branched off of the untested branch, and so those same warnings (as seen in the test blame and config files) still apply (mainly that it is untested).